Refactor : rename templateModal to TemplateModal #105
